requirements.txt
aiosmtplib==1.1.6
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.8
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
Jinja2==3.0.1
MarkupSafe==2.0.1
messages==0.8.0
python-dotenv==0.18.0
rfc3986==1.5.0
sniffio==1.2.0
validus==0.3.0
